Rename gradle plugins
