+++ gradle-plugins/src/main/kotlin/com/gchristov/thecodinglove/gradleplugins/backend/BackendNodeExecutablePlugin.kt
@@ -1,0 +1,19 @@
+package com.gchristov.thecodinglove.gradleplugins.backend
+
+import org.gradle.api.Plugin
+import org.gradle.api.Project
+import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
+
+@Suppress("unused")
+class BackendNodeExecutablePlugin : Plugin<Project> {
+    override fun apply(target: Project) {
+        target.plugins.apply("org.jetbrains.kotlin.multiplatform")
+        target.plugins.apply("dev.petuska.npm.publish")
+        target.extensions.configure(KotlinMultiplatformExtension::class.java) {
+            js(IR) {
+                binaries.library()
+                nodejs()
+            }
+        }
+    }
+}
