+++ gradle-plugins/src/main/kotlin/com/gchristov/thecodinglove/gradleplugins/web/WebBrowserExecutablePlugin.kt
@@ -1,11 +1,11 @@
-package com.gchristov.thecodinglove.gradleplugins.js
+package com.gchristov.thecodinglove.gradleplugins.web
 
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
 
 @Suppress("unused")
-class JavascriptBrowserExecutablePlugin : Plugin<Project> {
+class WebBrowserExecutablePlugin : Plugin<Project> {
     override fun apply(target: Project) {
         target.plugins.apply("org.jetbrains.kotlin.multiplatform")
         target.extensions.configure(KotlinMultiplatformExtension::class.java) {
@@ -16,20 +16,6 @@ class JavascriptBrowserExecutablePlugin : Plugin<Project> {
                         cssSupport.enabled = true
                     }
                 }
-            }
-        }
-    }
-}
-
-@Suppress("unused")
-class JavascriptNodeExecutablePlugin : Plugin<Project> {
-    override fun apply(target: Project) {
-        target.plugins.apply("org.jetbrains.kotlin.multiplatform")
-        target.plugins.apply("dev.petuska.npm.publish")
-        target.extensions.configure(KotlinMultiplatformExtension::class.java) {
-            js(IR) {
-                binaries.library()
-                nodejs()
             }
         }
     }
